adding more samples + built-in field editors
