+++ docs/static/playground/samples/field-editors/dropdowns/sample.js
@@ -2,31 +2,13 @@
 //% color="#FFAB19"
 namespace control {
 
-    /**
-     * Number dropdown
-     */
-
-    /**
-      * Get the time field editor
-      * @param ms time duration in milliseconds, eg: 500, 1000
-      */
-    //% blockId=timePicker block="%ms"
-    //% blockHidden=true
-    //% colorSecondary="#FFFFFF"
-    //% ms.fieldEditor="numberdropdown" ms.fieldOptions.decompileLiterals=true
-    //% ms.fieldOptions.data='[["100 ms", 100], ["200 ms", 200], ["500 ms", 500], ["1 second", 1000], ["2 seconds", 2000]]'
-    export function __timePicker(ms: number): number{
-        return ms;
-    }
-    
-
     //% block="pause %ms=timePicker"
     export function pause(ms: number) {
 
     }
 
     /**
-     * Text dropdown
+     * Custom text dropdown
      */
 
     /**
