+++ docs/static/playground/samples/field-editors/protractor/sample.js
@@ -8,17 +8,4 @@ namespace turtle {
     export function turn(angle: number) {
 
     }
-
-    /**
-     * A field editor that displays a protractor
-     */
-    //% blockId=protractorPicker block="%angle"
-    //% shim=TD_ID
-    //% angle.fieldEditor=protractor
-    //% angle.fieldOptions.decompileLiterals=1    
-    //% colorSecondary="#FFFFFF"
-    //% blockHidden=1
-    export function __protractor(angle: number) {
-        return angle;
-    }
 }
