+++ docs/static/playground/samples/field-editors/toggle/sample.js
@@ -3,9 +3,7 @@ namespace sample {
     /**
      * Render a boolean as a down/up toggle
      */
-    //% block="%down"
-    //% down.fieldEditor=toggledownup
-    //% down.fieldOptions.decompileLiterals=true
+    //% block="%down=toggleDownUp"
     export function downUp(down: boolean): boolean {
         return down;
     }
@@ -13,9 +11,7 @@ namespace sample {
     /**
      * Render a boolean as an up/down toggle
      */
-    //% block="%up"
-    //% up.fieldEditor=toggleupdown
-    //% up.fieldOptions.decompileLiterals=true
+    //% block="%up=toggleUpDown"
     export function upDown(up: boolean): boolean {
         return up;
     }
@@ -23,9 +19,7 @@ namespace sample {
     /**
      * Render a boolean as a high/low toggle
      */
-    //% block="%high"
-    //% high.fieldEditor=togglehighlow
-    //% high.fieldOptions.decompileLiterals=true
+    //% block="%high=toggleHighLow"
     export function highLow(high: boolean): boolean {
         return high;
     }
@@ -34,9 +28,7 @@ namespace sample {
     /**
      * Render a boolean as a on/off toggle
      */
-    //%  block="%on"
-    //% on.fieldEditor=toggleonoff
-    //% on.fieldOptions.decompileLiterals=true
+    //%  block="%on=toggleOnOff"
     export function onOff(on: boolean): boolean {
         return on;
     }
@@ -44,9 +36,7 @@ namespace sample {
     /**
      * Render a boolean as a yes/no toggle
      */
-    //% block="%yes"
-    //% yes.fieldEditor=toggleyesno
-    //% yes.fieldOptions.decompileLiterals=true
+    //% block="%yes=toggleYesNo"
     export function yesNo(yes: boolean): boolean {
         return yes;
     }
