+++ docs/static/playground/samples/field-editors/turnratio/sample.js
@@ -6,16 +6,4 @@ namespace motors {
     //% turnRatio.min=-200 turnRatio=200
     export function steer(turnRatio: number) {
     }
-
-
-    /**
-     * A turn ratio picker
-     * @param turnratio the turn ratio, eg: 0
-     */
-    //% blockId=turnRatioPicker block="%turnratio" shim=TD_ID
-    //% turnratio.fieldEditor="turnratio" colorSecondary="#FFFFFF"
-    //% blockHidden=1 turnRatio.fieldOptions.decompileLiterals=1
-    export function __turnRatioPicker(turnratio: number): number {
-        return turnratio;
-    }
 }
